agent plugin: report each btrfs fs only once
Often btrfs filesystems are mounted multiple times, especially when
subvolumes are used. In that case each mount point appears in
/proc/mounts with the same device but different mount points.

The scrub statistics are per device, though, not per
subvolume. Therefore it doesn't make sense to list a device more than
once.

With this change only the first mount point for each device will be
listed.